+++ src/main/java/ru/dm/shop/repository/ProductRepository.java
@@ -10,6 +10,7 @@ import org.springframework.transaction.annotation.Transactional;
 import ru.dm.shop.entity.Product;
 import ru.dm.shop.entity.ProductGroup;
 
+import java.math.BigDecimal;
 import java.util.Collection;
 import java.util.List;
 
@@ -33,7 +34,7 @@ public interface ProductRepository extends JpaRepository<Product, Long> {
     @Modifying
     @Transactional
     @Query("UPDATE Product product SET product.name = ?2, product.stock = ?3, product.retailPrice = ?4, product.wholesalePrice = ?5, product.currency = ?6, product.productGroup = ?7 where product.articule = ?1")
-    void updateProductByArticule(String articule, String name, int stock, float retailPrice, float wholesalePrice, String currency, ProductGroup productGroup);
+    void updateProductByArticule(String articule, String name, Integer stock, BigDecimal retailPrice, BigDecimal wholesalePrice, String currency, ProductGroup productGroup);
 
     Page<Product> findByProductGroupInAndStockGreaterThanAndNameContainingIgnoreCaseOrProductGroupInAndStockGreaterThanAndArticuleContainingIgnoreCase(Collection<ProductGroup> groups, int num, String name, Collection<ProductGroup> groups2, int num2, String articule, Pageable pageable);
     Page<Product> findByStockGreaterThanAndNameContainingIgnoreCase(int num, String name, Pageable pageable);
